+++ components/pages/todo.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { FormEvent, useEffect, useState } from "react";
-import { TodoListProps } from "@/types/Item";
+import { TodoItemProps, TodoListProps } from "@/types/Item";
 import TodoHeader from "@/components/organisms/todo/TodoHeader";
 import TodoList from "@/components/organisms/todo/TodoList";
 import styles from "@/components/organisms/todo/Todo.module.css";
@@ -11,16 +11,20 @@ function TodoPage() {
   const [todos, setTodos] = useState<TodoListProps>([]);
 
   useEffect(() => {
-    setTodos(() => {
-      const data = localStorage.getItem("data");
-
-      if (data) {
-        return JSON.parse(data);
+    const savedData = localStorage.getItem("data");
+    if (savedData) {
+      try {
+        setTodos(JSON.parse(savedData));
+      } catch (error) {
+        console.error("Failed to parse localStorage data:", error);
+        setTodos([]); // 기본값으로 초기화
       }
-
-      return [];
-    });
+    }
   }, []);
+
+  useEffect(() => {
+    if (todos.length) localStorage.setItem("data", JSON.stringify(todos));
+  }, [todos]);
 
   const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -39,7 +43,6 @@ function TodoPage() {
     };
 
     setTodos((prevState) => {
-      // 로컬 스토리지에 저장
       localStorage.setItem("data", JSON.stringify([...prevState, newTodo]));
 
       return [...prevState, newTodo];
@@ -48,12 +51,28 @@ function TodoPage() {
     e.currentTarget.reset();
   };
 
+  const handleTodos = (id: string, mode: TodoItemProps["status"]) => {
+    setTodos((prevState) =>
+      prevState.map((todo) => {
+        if (todo.id === id) {
+          return { ...todo, status: mode };
+        } else {
+          return todo;
+        }
+      })
+    );
+  };
+
   return (
     <div className={classNames(styles.TodoWrapper, "w-[400px]")}>
       <TodoHeader handleSubmit={handleSubmit} />
       {/* TodoList Area */}
       <div className="flex flex-row gap-12">
-        <TodoList todos={todos} handleSubmit={handleSubmit} />
+        <TodoList
+          todos={todos}
+          handleSubmit={handleSubmit}
+          handleTodos={handleTodos}
+        />
       </div>
     </div>
   );